backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
requests>=2.31.0
google-generativeai>=0.8.0
scikit-learn>=1.0.2
opencv-python>=4.5.5
matplotlib>=3.5.1
tqdm>=4.64.0
psutil>=5.9.0
pandas>=1.4.2
seaborn>=0.11.2
numpy>=1.22.3
pillow>=9.1.0
joblib>=1.1.0
xgboost>=1.6.0

